admin: Move Selection, SelectionApi, SelectionRoute, Selected back to src/ root
These files do not belong inside the editDialog/ subfolder since they are
general-purpose selection utilities not tightly coupled to EditDialog.
